+++ tests/test_commands.py
@@ -1,11 +1,181 @@
+import re
 from textwrap import dedent
 
-from ravenpy.config.commands import HRUState, HRUStateVariableTableCommand
+import pytest
+
+from ravenpy.config import options
+from ravenpy.config.commands import (
+    LU,
+    PL,
+    SOIL,
+    AdiabaticLapseRate,
+    EvaluationMetrics,
+    HRUState,
+    HRUStateVariableTableCommand,
+    LandUseClassesCommand,
+    LandUseParameterListCommand,
+    LinearTransform,
+    PotentialMeltMethod,
+    RunName,
+    SoilClassesCommand,
+    SoilParameterListCommand,
+    SoilProfilesCommand,
+    SuppressOutput,
+    VegetationClassesCommand,
+    VegetationParameterListCommand,
+)
+
+
+def test_adiabatic_lapse_rate_coefficient():
+    alr = AdiabaticLapseRate(0.2)
+    assert alr.to_rv() == ":AdiabaticLapseRate   0.2\n"
+
+
+def test_potential_melt_method_option():
+    expected = ":PotentialMeltMethod  POTMELT_EB\n"
+    assert PotentialMeltMethod("POTMELT_EB").to_rv() == expected
+    assert PotentialMeltMethod(options.PotentialMeltMethod.EB).to_rv() == expected
+
+
+def test_run_name():
+    assert RunName("test").to_rv() == ":RunName              test\n"
+
+
+def test_suppress_output():
+    assert SuppressOutput(True).to_rv() == ":SuppressOutput\n"
+    assert SuppressOutput(False).to_rv() == ""
 
 
 def test_hru_state():
     s = HRUState(index=1, data={"SOIL[0]": 1, "SOIL[1]": 2.0})
     assert s.to_rv() == "1,1.0,2.0"
+
+
+def test_evaluation_metrics():
+    em = EvaluationMetrics(["RMSE", "NASH_SUTCLIFFE"])
+    assert em.to_rv() == ":EvaluationMetrics    RMSE, NASH_SUTCLIFFE\n"
+
+
+def test_linear_transform():
+    assert LinearTransform(1, 0).to_rv() == ""
+
+
+def test_soil_classes():
+    c = SoilClassesCommand(
+        soil_classes=[
+            SoilClassesCommand.Record(n) for n in ["TOPSOIL", "FAST_RES", "SLOW_RES"]
+        ]
+    )
+    assert dedent(c.to_rv()) == dedent(
+        """
+    :SoilClasses
+        TOPSOIL
+        FAST_RES
+        SLOW_RES
+    :EndSoilClasses
+    """
+    )
+
+
+def test_vegetation_classes():
+    c = VegetationClassesCommand(
+        vegetation_classes=[VegetationClassesCommand.Record("VEG_ALL", 25, 6.0, 5.3)]
+    )
+    assert dedent(c.to_rv()) == dedent(
+        """
+    :VegetationClasses
+        :Attributes     ,        MAX_HT,       MAX_LAI, MAX_LEAF_COND
+        :Units          ,             m,          none,      mm_per_s
+        VEG_ALL         ,          25.0,           6.0,           5.3
+    :EndVegetationClasses
+    """
+    )
+
+
+def test_land_use_classe():
+    c = LandUseClassesCommand(land_use_classes=[LU("LU_ALL", 0, 1)])
+    assert dedent(c.to_rv()) == dedent(
+        """
+    :LandUseClasses
+        :Attributes     ,IMPERMEABLE_FRAC, FOREST_COVERAGE
+        :Units          ,           fract,           fract
+        LU_ALL          ,             0.0,             1.0
+    :EndLandUseClasses
+    """
+    )
+
+
+def test_soil_profiles(x=42):
+    c = SoilProfilesCommand(
+        soil_profiles=[
+            SOIL("DEFAULT_P", ["TOPSOIL", "FAST_RES", "SLOW_RES"], [x, 100.0, 100.0])
+        ]
+    )
+
+    pat = r"DEFAULT_P\s*,\s*3,\s*TOPSOIL,\s*(.+),\s*FAST_RES,\s*100.0,\s*SLOW_RES,\s*100.0"
+
+    m = re.findall(pat, c.to_rv())[0]
+    assert m == "42.0"
+
+
+def test_soil_parameter_list():
+    c = SoilParameterListCommand(
+        names=["POROSITY", "FIELD_CAPACITY"],
+        records=[
+            PL(name="[DEFAULT]", vals=[1, 0]),
+            PL(name="FAST_RES", vals=[None, None]),
+        ],
+    )
+
+    assert dedent(c.to_rv()) == dedent(
+        """
+    :SoilParameterList
+        :Parameters     ,          POROSITY,    FIELD_CAPACITY
+        :Units          ,              none,              none
+        [DEFAULT]       ,               1.0,               0.0
+        FAST_RES        ,          _DEFAULT,          _DEFAULT
+    :EndSoilParameterList
+    """
+    )
+
+
+def test_vegetation_parameter_list():
+    c = VegetationParameterListCommand(
+        names=["MAX_CAPACITY", "MAX_SNOW_CAPACITY", "RAIN_ICEPT_PCT", "SNOW_ICEPT_PCT"],
+        records=[PL(name="VEG_ALL", vals=[10000, 10000, 0.88, 0.88])],
+    )
+
+    assert dedent(c.to_rv()) == dedent(
+        """
+    :VegetationParameterList
+        :Parameters     ,      MAX_CAPACITY, MAX_SNOW_CAPACITY,    RAIN_ICEPT_PCT,    SNOW_ICEPT_PCT
+        :Units          ,              none,              none,              none,              none
+        VEG_ALL         ,           10000.0,           10000.0,              0.88,              0.88
+    :EndVegetationParameterList
+    """
+    )
+
+
+def test_land_use_parameter_list():
+    c = LandUseParameterListCommand(
+        names=["MELT_FACTOR", "MIN_MELT_FACTOR"],
+        records=[PL(name="[DEFAULT]", vals=[1, 2.2])],
+    )
+    assert dedent(c.to_rv()) == dedent(
+        """
+    :LandUseParameterList
+        :Parameters     ,       MELT_FACTOR,   MIN_MELT_FACTOR
+        :Units          ,              none,              none
+        [DEFAULT]       ,               1.0,               2.2
+    :EndLandUseParameterList
+    """
+    )
+
+    with pytest.raises(ValueError):
+        LandUseParameterListCommand(
+            names=["MELT_FACTOR", "MIN_MELT_FACTOR"],
+            records=[PL(name="[DEFAULT]", vals=[None, 2.2])],
+        )
 
 
 class TestHRUStateVariableTableCommand: